# docs/swagger.yaml
basePath: /v1
definitions:
  model.Card:
    properties:
      activated:
        description: Activated the cart should be activated before usage.
        type: boolean
      container_uuid:
        type: string
      creation_date:
        type: string
      pic:
        description: Pic is the uri of the image source link.
        type: string
      url:
        type: string
      uuid:
        type: string
    type: object
  model.Contener:
    properties:
      cards omitempty:
        items:
          $ref: '#/definitions/model.Card'
        type: array
      creation_date:
        type: string
      secret:
        type: string
      title:
        type: string
      user_id:
        type: string
      uuid:
        type: string
    type: object
  model.User:
    properties:
      creation_date:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      uuid:
        type: string
    type: object
  model.UserLogin:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is an API for creating hash in order to create keypasses.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger For Keypass API
  version: "1.0"
paths:
  /cards:
    get:
      consumes:
      - application/json
      description: get a Card by ID
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/model.Card'
            type: array
        "400":
          description: We need ID!!
          schema:
            type: string
        "404":
          description: Can not find ID
          schema:
            type: string
      security:
      - BearerAuth: []
    patch:
      consumes:
      - application/json
      description: update a Card from the payload.
      parameters:
      - description: Add a Card
        in: body
        name: Card
        required: true
        schema:
          $ref: '#/definitions/model.Card'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Card'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BearerAuth: []
    post:
      consumes:
      - application/json
      description: create a Card from the payload.
      parameters:
      - description: Add a Card
        in: body
        name: Card
        required: true
        schema:
          $ref: '#/definitions/model.Card'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Card'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BearerAuth: []
  /cards/{uuid}:
    delete:
      consumes:
      - application/json
      description: delete a Card by ID
      parameters:
      - description: Some ID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/model.Card'
        "400":
          description: We need ID!!
          schema:
            type: string
        "404":
          description: Can not find ID
          schema:
            type: string
      security:
      - BearerAuth: []
    get:
      consumes:
      - application/json
      description: get a Card by ID
      parameters:
      - description: Some ID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/model.Card'
        "400":
          description: We need ID!!
          schema:
            type: string
        "404":
          description: Can not find ID
          schema:
            type: string
      security:
      - BearerAuth: []
  /conteners:
    get:
      consumes:
      - application/json
      description: get a Contener by ID
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/model.Contener'
            type: array
        "400":
          description: We need ID!!
          schema:
            type: string
        "404":
          description: Can not find ID
          schema:
            type: string
      security:
      - BearerAuth: []
    patch:
      consumes:
      - application/json
      description: update a Contener from the payload.
      parameters:
      - description: Add a Contener
        in: body
        name: Contener
        required: true
        schema:
          $ref: '#/definitions/model.Contener'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Contener'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BearerAuth: []
    post:
      consumes:
      - application/json
      description: create a Contener from the payload.
      parameters:
      - description: Add a Contener
        in: body
        name: Contener
        required: true
        schema:
          $ref: '#/definitions/model.Contener'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Contener'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BearerAuth: []
  /conteners/{uuid}:
    delete:
      consumes:
      - application/json
      description: delete a Contener by ID
      parameters:
      - description: Some ID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/model.Contener'
        "400":
          description: We need ID!!
          schema:
            type: string
        "404":
          description: Can not find ID
          schema:
            type: string
      security:
      - BearerAuth: []
    get:
      consumes:
      - application/json
      description: get a Contener by ID
      parameters:
      - description: Some ID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/model.Contener'
        "400":
          description: We need ID!!
          schema:
            type: string
        "404":
          description: Can not find ID
          schema:
            type: string
      security:
      - BearerAuth: []
  /login:
    post:
      consumes:
      - application/json
      description: check if the login pass are correct and gives backe a JWT value
      parameters:
      - description: Add a User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
  /users:
    get:
      consumes:
      - application/json
      description: get a User by ID
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
        "400":
          description: We need ID!!
          schema:
            type: string
        "404":
          description: Can not find ID
          schema:
            type: string
    patch:
      consumes:
      - application/json
      description: update a User from the payload.
      parameters:
      - description: Add a User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            type: string
    post:
      consumes:
      - application/json
      description: create a User from the payload.
      parameters:
      - description: Add a User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            type: string
  /users/{uuid}:
    delete:
      consumes:
      - application/json
      description: delete a User by ID
      parameters:
      - description: Some ID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: We need ID!!
          schema:
            type: string
        "404":
          description: Can not find ID
          schema:
            type: string
    get:
      consumes:
      - application/json
      description: get a User by ID
      parameters:
      - description: Some ID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: We need ID!!
          schema:
            type: string
        "404":
          description: Can not find ID
          schema:
            type: string
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"